requirements.txt
django==3.2
djangorestframework==3.13.1
django-cors-headers==3.13.1
channels==3.0.5
mysqlclient==2.1.1